+++ components/collector/src/source_collectors/axe_csv.py
@@ -2,8 +2,9 @@
 
 import csv
 import hashlib
-from io import StringIO
 import re
+from io import StringIO
+from typing import Dict, List, Tuple
 
 from utilities.type import Responses, Value, Entities
 from .source_collector import SourceCollector
@@ -20,13 +21,16 @@ class AxeCSVAccessibility(SourceCollector):
         """Convert csv rows of the Axe report into entities, in this case accessibility violations."""
         return [
             dict(
-                key=hashlib.md5(str(row).encode('utf-8')).hexdigest(),  # nosec
-                violation_type=row["Violation Type"], impact=row["Impact"], url=str(row["URL"]),
-                element=row["DOM Element"], page=re.sub(r'http[s]?://[^/]+', '', row['URL']),
-                description=row["Messages"], help=row["Help"])
-            for row in self.__parse_csv(responses)]
+                key=hashlib.md5(row.encode('utf-8')).hexdigest(),  # nosec
+                violation_type=items["Violation Type"], impact=items["Impact"], url=str(items["URL"]),
+                element=items["DOM Element"], page=re.sub(r'http[s]?://[^/]+', '', items['URL']),
+                description=items["Messages"], help=items["Help"])
+            for row, items in self.__parse_csv(responses)]
 
-    @staticmethod
-    def __parse_csv(responses: Responses) -> csv.DictReader:
-        """Parse the CSV and return the row iterator."""
-        return csv.DictReader(StringIO(responses[0].text, newline=None))
+    def __parse_csv(self, responses: Responses) -> List[Tuple[str, Dict[str, str]]]:
+        """Parse the CSV and return the rows and parsed items ."""
+        impact_levels = self._parameter("impact")
+        rows = responses[0].text.split("\n")[1:]
+        parsed_rows = csv.DictReader(StringIO(responses[0].text, newline=None))
+        return [
+            (row, parsed_row) for row, parsed_row in zip(rows, parsed_rows) if parsed_row["Impact"] in impact_levels]